backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pymysql==1.1.0
pymongo==4.6.1
cryptography==42.0.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
stripe==7.11.0
redis==5.0.1
pika==1.3.2
bleach==6.1.0
python-multipart==0.0.6
email-validator==2.1.0
celery==5.3.4
user-agents==2.2.0
jinja2==3.1.2
